+++ README.md
@@ -1,0 +1,2 @@
+# Team_DragonSeller
+21년 서울과학기술대학교 소프트웨어공학 팀프로젝트 깃허브입니다.
