+++ package.json
@@ -18,11 +18,14 @@
   },
   "homepage": "https://github.com/SoinDosa/Team_DragonSeller#readme",
   "dependencies": {
+    "axios": "^0.19.2",
     "bcrypt": "^5.0.1",
     "body-parser": "^1.19.0",
     "cookie-parser": "^1.4.5",
     "express": "^4.17.1",
     "jsonwebtoken": "^8.5.1",
-    "mongoose": "^5.12.3"
+    "mongoose": "^5.12.3",
+    "multer": "^1.4.2",
+    "react-dropzone": "^11.3.2"
   }
 }
