--- server/config/dev.js
@@ -1,3 +1,0 @@
-module.exports = {
-    mongoURI: 'mongodb+srv://sunny:151515@ds.mc5fr.mongodb.net/myFirstDatabase?retryWrites=true&w=majority'
-}
